[Validator] refactored into own package.
